requirements.txt
aiohttp==3.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
beautifulsoup4==4.6.0
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3
dateutils==0.6.6
discord.py==1.0.0a1481+g983be10
idna==2.7
idna-ssl==1.0.1
multidict==4.3.1
numpy==1.14.5
Pillow==5.1.0
pycparser==2.18
PyMySQL==0.9.0
PyNaCl==1.1.2
python-dateutil==2.7.3
pytz==2018.5
pyzbar==0.1.7
qrcode==6.0
six==1.11.0
SQLAlchemy==1.2.9
websockets==5.0.1
yarl==1.2.6
